move create_invoices to a service dir
