+++ pkg/handlers/internalapi/shipments_test.go
@@ -4,7 +4,6 @@ import (
 	"net/http/httptest"
 	"time"
 
-	"github.com/facebookgo/clock"
 	"github.com/go-openapi/strfmt"
 	"github.com/go-openapi/swag"
 	shipmentop "github.com/transcom/mymove/pkg/gen/internalapi/internaloperations/shipments"
@@ -337,23 +336,6 @@ func (suite *HandlerSuite) TestCompleteHHGHandler() {
 	suite.Assertions.IsType(&shipmentop.CompleteHHGOK{}, response)
 	okResponse := response.(*shipmentop.CompleteHHGOK)
 	suite.Equal("COMPLETED", string(okResponse.Payload.Status))
-}
-
-func (suite *HandlerSuite) TestCreateInvoicesCall() {
-	shipmentLineItem := testdatagen.MakeDefaultShipmentLineItem(suite.TestDB())
-	suite.TestDB().Eager("ShipmentLineItems.ID").Reload(&shipmentLineItem.Shipment)
-
-	createInvoices := CreateInvoices{
-		suite.TestDB(),
-		[]models.Shipment{shipmentLineItem.Shipment},
-	}
-	invoices, err := createInvoices.Call(clock.NewMock())
-	suite.NoError(err)
-
-	suite.Equal(1, len(invoices))
-	suite.Equal(models.InvoiceStatusINPROCESS, invoices[0].Status)
-	suite.Equal(1, len(invoices[0].ShipmentLineItems))
-	suite.Equal(invoices[0].ID, *invoices[0].ShipmentLineItems[0].InvoiceID)
 }
 
 /*
